[sdk] gating lib & GatingInspector v1
